+++ src/controller/mod.rs
@@ -1,0 +1,2 @@
+pub mod articles;
+pub mod top;
