+++ src/controller/top.rs
@@ -1,0 +1,4 @@
+#[get("/")]
+pub fn top() -> String {
+  "Hello, world!".to_string()
+}
