+++ src/lib.rs
@@ -1,0 +1,7 @@
+#![feature(proc_macro_hygiene, decl_macro)]
+
+#[macro_use]
+extern crate rocket;
+
+pub mod controller;
+pub mod routes;
